Refactor plot helpers into MatrixPlot
