+++ libplot/MatrixPlot.h
@@ -1,7 +1,16 @@
 #ifndef MATRIXPLOT_H
 #define MATRIXPLOT_H
 
+#include <algorithm>
 #include <string>
+#include <utility>
+#include <vector>
+
+#include "AnalysisDataLoader.h"
+#include "AnalysisResult.h"
+#include "HistogramCut.h"
+#include "QuadTreeBinning2D.h"
+#include "Selection.h"
 
 #include "TCanvas.h"
 #include "TH2F.h"
@@ -12,35 +21,207 @@
 namespace analysis {
 
 class MatrixPlot : public IHistogramPlot {
-  public:
-    MatrixPlot(std::string plot_name, TH2F *hist, std::string output_directory = "plots")
-        : IHistogramPlot(std::move(plot_name), std::move(output_directory)), hist_(hist) {}
+public:
+  MatrixPlot(std::string plot_name, const VariableResult &x_res,
+             const VariableResult &y_res, AnalysisDataLoader &loader,
+             const Selection &selection, std::string output_directory = "plots",
+             const std::vector<Cut> &x_cuts = {},
+             const std::vector<Cut> &y_cuts = {})
+      : IHistogramPlot(std::move(plot_name), std::move(output_directory)) {
+    auto edges = determineEdges(loader, x_res, y_res);
+    hist_ = new TH2F(plot_name_.c_str(), plot_name_.c_str(),
+                     edges.first.size() - 1, edges.first.data(),
+                     edges.second.size() - 1, edges.second.data());
+    hist_->GetXaxis()->SetTitle(x_res.binning_.getTexLabel().c_str());
+    hist_->GetYaxis()->SetTitle(y_res.binning_.getTexLabel().c_str());
 
-    ~MatrixPlot() override { delete hist_; }
+    std::string filter = selection.str();
 
-  private:
-    void draw(TCanvas &canvas) override {
-        canvas.cd();
-        const int stats_off = 0;
-        const int contour_count = 255;
-        const double margin = 0.15;
-        const double title_offset = 1.2;
+    for (auto &kv : loader.getSampleFrames()) {
+      auto df = kv.second.nominal_node_;
+      df = applySelectionFilters(df, x_res, y_res, filter, x_cuts, y_cuts);
+      auto data = extractValues(df, x_res, y_res);
+      fillHistogram(hist_, data);
+    }
+  }
 
-        gStyle->SetOptStat(stats_off);
-        gStyle->SetNumberContours(contour_count);
+  ~MatrixPlot() override { delete hist_; }
 
-        canvas.SetLogz();
-        canvas.SetLeftMargin(margin);
-        canvas.SetRightMargin(margin);
+private:
+  static std::pair<std::vector<double>, std::vector<double>>
+  determineEdges(AnalysisDataLoader &loader, const VariableResult &x_res,
+                 const VariableResult &y_res) {
+    auto x_edges = x_res.binning_.getEdges();
+    auto y_edges = y_res.binning_.getEdges();
 
-        hist_->SetTitle("");
-        hist_->GetZaxis()->SetTitleOffset(title_offset);
-        hist_->Draw("COLZ");
+    std::vector<ROOT::RDF::RNode> mc_nodes;
+    for (auto &[key, sample] : loader.getSampleFrames())
+      if (sample.isMc())
+        mc_nodes.emplace_back(sample.nominal_node_);
+    if (!mc_nodes.empty()) {
+      auto bins = QuadTreeBinning2D::calculate(mc_nodes, x_res.binning_,
+                                               y_res.binning_);
+      x_edges = bins.first.getEdges();
+      y_edges = bins.second.getEdges();
     }
 
-    TH2F *hist_;
+    return {x_edges, y_edges};
+  }
+
+  static ROOT::RDF::RNode
+  applySelectionFilters(ROOT::RDF::RNode df, const VariableResult &x_res,
+                        const VariableResult &y_res, const std::string &filter,
+                        const std::vector<Cut> &x_cuts,
+                        const std::vector<Cut> &y_cuts) {
+    if (filter.find_first_not_of(" \t\n\r") != std::string::npos)
+      df = df.Filter(filter);
+
+    for (auto const &c : x_cuts) {
+      std::string expr =
+          x_res.binning_.getVariable() +
+          (c.direction == CutDirection::GreaterThan ? ">" : "<") +
+          std::to_string(c.threshold);
+      df = df.Filter(expr);
+    }
+
+    for (auto const &c : y_cuts) {
+      std::string expr =
+          y_res.binning_.getVariable() +
+          (c.direction == CutDirection::GreaterThan ? ">" : "<") +
+          std::to_string(c.threshold);
+      df = df.Filter(expr);
+    }
+
+    return df;
+  }
+
+  struct SampleData {
+    bool x_is_vec;
+    bool y_is_vec;
+    std::vector<double> ws;
+    std::vector<double> xs;
+    std::vector<double> ys;
+    std::vector<std::vector<double>> xs_vec;
+    std::vector<std::vector<double>> ys_vec;
+  };
+
+  static SampleData extractValues(ROOT::RDF::RNode df,
+                                  const VariableResult &x_res,
+                                  const VariableResult &y_res) {
+    SampleData data;
+
+    bool has_w = df.HasColumn("nominal_event_weight");
+    const auto &x_col = x_res.binning_.getVariable();
+    const auto &y_col = y_res.binning_.getVariable();
+
+    auto x_type = df.GetColumnType(x_col);
+    auto y_type = df.GetColumnType(y_col);
+    data.x_is_vec = x_type.find("vector") != std::string::npos ||
+                    x_type.find("RVec") != std::string::npos;
+    data.y_is_vec = y_type.find("vector") != std::string::npos ||
+                    y_type.find("RVec") != std::string::npos;
+
+    size_t n_events = df.Count().GetValue();
+    if (has_w) {
+      auto w_type = df.GetColumnType("nominal_event_weight");
+      if (w_type.find("float") != std::string::npos) {
+        auto wv = df.Take<float>("nominal_event_weight").GetValue();
+        data.ws.assign(wv.begin(), wv.end());
+      } else {
+        data.ws = df.Take<double>("nominal_event_weight").GetValue();
+      }
+    } else {
+      data.ws.assign(n_events, 1.0);
+    }
+
+    auto get_scalar = [&](const std::string &col, const std::string &type) {
+      std::vector<double> vals;
+      if (type.find("float") != std::string::npos) {
+        auto v = df.Take<float>(col).GetValue();
+        vals.assign(v.begin(), v.end());
+      } else if (type.find("int") != std::string::npos ||
+                 type.find("unsigned") != std::string::npos) {
+        auto v = df.Take<int>(col).GetValue();
+        vals.assign(v.begin(), v.end());
+      } else {
+        vals = df.Take<double>(col).GetValue();
+      }
+      return vals;
+    };
+
+    auto get_vector = [&](const std::string &col, const std::string &type) {
+      std::vector<std::vector<double>> vals;
+      if (type.find("float") != std::string::npos) {
+        auto v = df.Take<std::vector<float>>(col).GetValue();
+        for (auto &inner : v)
+          vals.emplace_back(inner.begin(), inner.end());
+      } else if (type.find("int") != std::string::npos ||
+                 type.find("unsigned") != std::string::npos) {
+        auto v = df.Take<std::vector<int>>(col).GetValue();
+        for (auto &inner : v)
+          vals.emplace_back(inner.begin(), inner.end());
+      } else {
+        vals = df.Take<std::vector<double>>(col).GetValue();
+      }
+      return vals;
+    };
+
+    if (!data.x_is_vec)
+      data.xs = get_scalar(x_col, x_type);
+    else
+      data.xs_vec = get_vector(x_col, x_type);
+
+    if (!data.y_is_vec)
+      data.ys = get_scalar(y_col, y_type);
+    else
+      data.ys_vec = get_vector(y_col, y_type);
+
+    return data;
+  }
+
+  static void fillHistogram(TH2F *hist, const SampleData &data) {
+    if (!data.x_is_vec && !data.y_is_vec) {
+      for (size_t i = 0; i < data.xs.size(); ++i)
+        hist->Fill(data.xs[i], data.ys[i], data.ws[i]);
+    } else if (data.x_is_vec && !data.y_is_vec) {
+      for (size_t i = 0; i < data.xs_vec.size(); ++i)
+        for (double xv : data.xs_vec[i])
+          hist->Fill(xv, data.ys[i], data.ws[i]);
+    } else if (!data.x_is_vec && data.y_is_vec) {
+      for (size_t i = 0; i < data.ys_vec.size(); ++i)
+        for (double yv : data.ys_vec[i])
+          hist->Fill(data.xs[i], yv, data.ws[i]);
+    } else {
+      for (size_t i = 0; i < data.xs_vec.size(); ++i) {
+        size_t lim = std::min(data.xs_vec[i].size(), data.ys_vec[i].size());
+        for (size_t j = 0; j < lim; ++j)
+          hist->Fill(data.xs_vec[i][j], data.ys_vec[i][j], data.ws[i]);
+      }
+    }
+  }
+
+  void draw(TCanvas &canvas) override {
+    canvas.cd();
+    const int stats_off = 0;
+    const int contour_count = 255;
+    const double margin = 0.15;
+    const double title_offset = 1.2;
+
+    gStyle->SetOptStat(stats_off);
+    gStyle->SetNumberContours(contour_count);
+
+    canvas.SetLogz();
+    canvas.SetLeftMargin(margin);
+    canvas.SetRightMargin(margin);
+
+    hist_->SetTitle("");
+    hist_->GetZaxis()->SetTitleOffset(title_offset);
+    hist_->Draw("COLZ");
+  }
+
+  TH2F *hist_;
 };
 
-}
+} // namespace analysis
 
 #endif
